Rewrite draft email pages in React (#96)
* Add debugger, rewrite email draft page in React

* Get edit working, need email attachments

* Working on file uploads

* Finish email draft rebuild

+++ app/case/views/case/email.py
@@ -4,8 +4,7 @@ from typing import List
 from django.utils.text import slugify
 
 from django.http import Http404, HttpResponse
-from django.shortcuts import render, redirect
-from django.contrib import messages
+from django.shortcuts import redirect
 from django.views.decorators.http import require_http_methods
 from django.utils.html import strip_tags
 from html_sanitizer import Sanitizer
@@ -13,10 +12,12 @@ from django.urls import reverse
 from django.template.loader import render_to_string
 from django.utils.safestring import mark_safe
 from django.db.models import Q
+from django.db import transaction
+from django.core.files.base import ContentFile
 from bs4 import BeautifulSoup
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
-
+from rest_framework.exceptions import ValidationError
 
 from core.models import Issue
 from case.views.auth import paralegal_or_better_required
@@ -28,8 +29,6 @@ from emails.models import (
     EmailAttachment,
     SharepointState,
 )
-from case.forms import EmailForm
-from case.utils import merge_form_data
 from case.utils.router import Router
 from microsoft.endpoints import MSGraphAPI
 from microsoft.service import save_email_attachment
@@ -38,6 +37,7 @@ from case.serializers import (
     IssueDetailSerializer,
     EmailSerializer,
     EmailAttachmentSerializer,
+    EmailTemplateSerializer,
     EmailThreadSerializer,
 )
 from case.views.case.detail import get_detail_urls
@@ -179,119 +179,93 @@ def _process_email_for_display(email: Email):
 
 @router.use_route("draft")
 @paralegal_or_better_required
-@require_http_methods(["GET", "POST"])
+@api_view(["GET", "POST"])
 def email_draft_create_view(request, pk):
     issue = _get_issue_for_emails(request, pk)
-    case_email_address = build_clerk_address(issue, email_only=True)
-    case_emails = get_case_emails(issue)
-    templates = EmailTemplate.objects.filter(
-        Q(topic=issue.topic) | Q(topic="GENERAL")
-    ).order_by("-created_at")
-    for template in templates:
-        # Annotate with url
-        qs = request.GET.copy()
-        qs["template"] = template.pk
-        template.url = "?" + qs.urlencode()
-
-    parent_email = None
     if request.method == "POST":
-        default_data = {
-            "from_address": case_email_address,
+        data = {
+            "from_address": build_clerk_address(issue, email_only=True),
             "state": EmailState.DRAFT,
-            "issue": issue,
             "sender": request.user,
+            **request.data,
         }
-        data = merge_form_data(request.POST, default_data)
-        form = EmailForm(data, files=request.FILES)
-        if form.is_valid():
-            email = form.save()
-            messages.success(request, "Draft created")
-            return redirect("case-email-edit", issue.pk, email.pk)
+        serializer = EmailSerializer(data=data)
+        serializer.is_valid(raise_exception=True)
+        serializer.save(issue=issue)
+        return Response(serializer.data)
     else:
-        # It's a GET request.
+        templates = EmailTemplate.objects.filter(
+            Q(topic=issue.topic) | Q(topic="GENERAL")
+        ).order_by("created_at")
         parent_id = request.GET.get("parent")
-        template_id = request.GET.get("template")
-        initial = {}
+        parent_email = None
         if parent_id:
             try:
-                parent_email = Email.objects.get(id=parent_id)
-                _process_email_for_display(parent_email)
-                initial["subject"] = parent_email.subject
-                if parent_email.state == EmailState.INGESTED:
-                    initial["to_address"] = parent_email.from_address
-                else:
-                    initial["to_address"] = parent_email.to_address
+                parent_email = EmailSerializer(Email.objects.get(pk=parent_id)).data
             except Email.DoesNotExist:
                 raise Http404()
-        if template_id:
-            try:
-                template = EmailTemplate.objects.get(id=template_id)
-                initial["text"] = template.text
-                initial["subject"] = template.subject
-            except EmailTemplate.DoesNotExist:
-                raise Http404()
 
-        form = EmailForm(initial=initial)
-
-    context = {
-        "issue": issue,
-        "parent_email": parent_email,
-        "form": form,
-        "case_emails": case_emails,
-        "case_email_address": case_email_address,
-        "templates": templates,
-        "is_disabled": False,
-    }
-    return render(request, "case/case/email/draft_create.html", context)
+        context = {
+            "case_email_url": reverse("case-email-list", args=(issue.pk,)),
+            "parent_email": parent_email,
+            "issue": IssueDetailSerializer(issue).data,
+            "templates": EmailTemplateSerializer(templates, many=True).data,
+        }
+        return render_react_page(
+            request, f"Case {issue.fileref}", "email-draft-create", context
+        )
 
 
 @router.use_route("edit")
 @paralegal_or_better_required
-@require_http_methods(["GET", "POST", "DELETE"])
+@api_view(["GET", "PATCH", "DELETE"])
 def email_draft_edit_view(request, pk, email_pk):
     issue = _get_issue_for_emails(request, pk)
     email = _get_email_for_issue(issue, email_pk)
     if not email.state in DRAFT_EMAIL_STATES:
         return redirect("case-email-list", issue.pk)
-
-    case_email_address = build_clerk_address(issue, email_only=True)
-    case_emails = get_case_emails(issue)
-    if request.method == "POST":
-        default_data = {
-            "from_address": case_email_address,
-            "state": EmailState.DRAFT,
-            "issue": issue,
-            "sender": request.user,
-        }
-        data = merge_form_data(request.POST, default_data)
-        form = EmailForm(data, instance=email, files=request.FILES)
-        if form.is_valid():
-            messages.success(request, "Draft saved.")
-            email = form.save()
+    if request.method == "PATCH":
+        serializer = EmailSerializer(email, data=request.data, partial=True)
+        serializer.is_valid(raise_exception=True)
+        serializer.save()
+        return Response(serializer.data)
     elif request.method == "DELETE":
-        email.delete()
-        messages.success(request, "Draft deleted")
-        return HttpResponse(
-            headers={"HX-Redirect": reverse("case-email-list", args=(pk,))}
-        )
+        with transaction.atomic():
+            email.emailattachment_set.all().delete()
+            email.delete()
+        return Response({})
     else:
-        form = EmailForm(instance=email)
+        # GET request.
+        api = MSGraphAPI()
+        sharepoint_docs = api.folder.get_all_files(f"cases/{issue.id}")
+        context = {
+            "email_preview_url": reverse(
+                "case-email-preview", args=(issue.pk, email.pk)
+            ),
+            "case_email_url": reverse("case-email-list", args=(issue.pk,)),
+            "issue": IssueDetailSerializer(issue).data,
+            "email": EmailSerializer(email).data,
+            "sharepoint_docs": sharepoint_docs,
+        }
+        return render_react_page(
+            request, f"Case {issue.fileref}", "email-draft-edit", context
+        )
 
-    api = MSGraphAPI()
-    sharepoint_docs = api.folder.get_all_files(f"cases/{issue.id}")
-    sharepoint_options = [
-        {"name": doc["name"], "value": doc["id"]} for doc in sharepoint_docs
-    ]
-    context = {
-        "issue": issue,
-        "form": form,
-        "case_emails": case_emails,
-        "email": email,
-        "case_email_address": case_email_address,
-        "is_disabled": email.state != EmailState.DRAFT,
-        "sharepoint_options": sharepoint_options,
-    }
-    return render(request, "case/case/email/draft_edit.html", context)
+
+@router.use_route("send")
+@paralegal_or_better_required
+@api_view(["POST"])
+def email_draft_send_view(request, pk, email_pk):
+    issue = _get_issue_for_emails(request, pk)
+    email = _get_email_for_issue(issue, email_pk)
+    if email.state == EmailState.DRAFT:
+        email.state = EmailState.READY_TO_SEND
+        email.html = _render_email_template(email.html)
+        email.save()
+    else:
+        raise Http404()
+
+    return Response(status=200)
 
 
 @router.use_route("preview")
@@ -319,51 +293,49 @@ def _render_email_template(html):
         context = {"html": mark_safe(soup.body.decode_contents())}
     else:
         context = {"html": ""}
-    return render_to_string("case/case/email/preview.html", context)
-
-
-@router.use_route("send")
-@paralegal_or_better_required
-@require_http_methods(["POST"])
-def email_draft_send_view(request, pk, email_pk):
-    issue = _get_issue_for_emails(request, pk)
-    email = _get_email_for_issue(issue, email_pk)
-    if email.state == EmailState.DRAFT:
-        email.state = EmailState.READY_TO_SEND
-        email.html = _render_email_template(email.html)
-        email.save()
-    elif email.state == EmailState.SENT:
-        # Redirect to list view
-        messages.success(request, "Email sent")
-        return HttpResponse(
-            headers={"HX-Redirect": reverse("case-email-list", args=(pk,))}
-        )
-    elif email.state != EmailState.READY_TO_SEND:
-        raise Http404()
-
-    form = EmailForm(instance=email)
-    context = {
-        "issue": issue,
-        "form": form,
-        "email": email,
-        "is_disabled": True,
-    }
-    return render(request, "case/case/email/_email_form.html", context)
+    return render_to_string("case/case/email_preview.html", context)
 
 
 @router.use_route("attach")
 @paralegal_or_better_required
-@require_http_methods(["DELETE"])
-def email_attachment_view(request, pk, email_pk, attach_pk):
+@api_view(["POST", "DELETE"])
+def email_attachment_view(request, pk, email_pk, attach_pk=None):
     """Delete the email attachment, return HTML fragment for htmx."""
     issue = _get_issue_for_emails(request, pk)
     email = _get_email_for_issue(issue, email_pk)
     if not email.state == EmailState.DRAFT:
         raise Http404()
 
-    email.emailattachment_set.filter(id=attach_pk).delete()
-    context = {"issue": issue, "email": email}
-    return render(request, "case/case/email/_draft_attachments.html", context)
+    if request.method == "DELETE":
+        email.emailattachment_set.filter(id=attach_pk).delete()
+        return Response(status=200)
+
+    else:
+        # POST request.
+        if "sharepoint_id" in request.data:
+            # Download attachment from SharePoint.
+            sharepoint_id = request.data["sharepoint_id"]
+            api = MSGraphAPI()
+            filename, content_type, file_bytes = api.folder.download_file(sharepoint_id)
+            # Save as email attachment
+            f = ContentFile(file_bytes, name=filename)
+        elif "file" in request.data:
+            f = request.data["file"]
+            content_type = f.content_type
+        else:
+            msg = "Could not find and uploaded file or SharePoint document."
+            raise ValidationError(msg)
+
+        if f.size / 1024 / 1024 > 30:
+            raise ValidationError({"file": "File must be <30MB."})
+
+        attach = EmailAttachment.objects.create(
+            email=email,
+            file=f,
+            content_type=content_type,
+        )
+        serializer = EmailAttachmentSerializer(attach)
+        return Response(serializer.data)
 
 
 @router.use_route("attach-upload")
